Responsive website via ReactJS
Responsive website with ReactJS, non build.

To build: You can build by opening the file folder using CMD or the console and typing npm build or yarn build. In order to share this directly on your own site, it will be sufficient to open the build folder that will be created in the main directory after building and put the contents in the main directory.